Move to parcel to simplify everything.
