--- .eslintrc.js
@@ -1,3 +1,0 @@
-const neutrino = require('neutrino');
-
-module.exports = neutrino().eslintrc();
