+++ demo.html
@@ -1,0 +1,9 @@
+<html>
+<head>
+  <title>react-mathjax-review</title>
+</head>
+<body>
+  <div id="root"></div>
+  <script src="./demo.jsx"></script>
+</body>
+</html>
