--- jest.config.js
@@ -1,5 +1,0 @@
-const neutrino = require('neutrino');
-
-process.env.NODE_ENV = process.env.NODE_ENV || 'test';
-
-module.exports = neutrino().jest();
