--- src/App/App.css
@@ -1,3 +1,0 @@
-.App {
-  padding: 20px;
-}
