--- src/App/index.jsx
@@ -1,3 +1,0 @@
-import App from './App'
-
-export default App
